requirements.txt
# --- Web Server ---
Flask==3.0.3
requests==2.32.3
google-generativeai==0.7.2
python-dotenv==1.0.1

# --- Core API ---
fastapi==0.111.0
uvicorn[standard]==0.30.1

# --- Pydantic ---
pydantic==2.8.2
pydantic-core==2.20.1

# --- HTTP client (Ollama / OpenAI-compatible) ---
httpx==0.27.0

# --- Retrieval (BM25) ---
rank-bm25==0.2.2
numpy==1.26.4

# --- JSON tối ưu ---
orjson==3.11.3
ujson==5.11.0

# --- Tuỳ chọn ---
sentence-transformers==2.7.0
scikit-learn==1.4.2
faiss-cpu==1.8.0.post1
# torch==2.3.1
starlette==0.37.2
rich==13.7.1
